Improve type classification of static bodies
When checking if a body is static in the fluid discretization, the code
now checks if the body's frame is `DiscretizationFrame` or is the same
base frame as the fluid's discretization.

Before, the body has to explicitly have the `DiscretizationFrame` if it
was to be static in the fluid grid. Now, if both the fluid grid and
bodies are in the `GlobalFrame` or any other `BaseFrame`, the bodies
will be considered static.